bit::Set now has template parameter N

+++ include/dctl/bit/iterator/iterator_1.hpp
@@ -42,7 +42,7 @@ public:
         constexpr ConstIterator(Block const* b, U const& value)
         :
                 block_{b},
-                index_{value}
+                index_{static_cast<int>(value)}
         {
                 assert(b != nullptr);
                 assert(value == N);
